arreglo de onclick en unas boletas que se habia eliminado

+++ vistaagresiones.php
@@ -75,7 +75,7 @@
                                     </td>
                                     <td>
                                         <a href="editagresiones.php?id=<?php echo htmlspecialchars($row['id'], ENT_QUOTES, 'UTF-8'); ?>" title="Click para editar la boleta agresiones"><i class="fa fa-edit fa-lg text-warning"></i></a>
-                                        <a href="deleteagresiones.php?id=<?php echo htmlspecialchars($row['id'], ENT_QUOTES, 'UTF-8'); ?>" title="Click para eliminar la boleta agresiones"><i class="fa fa-trash fa-lg text-danger"></i></a>
+                                        <a href="deleteagresiones.php?id=<?php echo htmlspecialchars($row['id'], ENT_QUOTES, 'UTF-8'); ?>" title="Click para eliminar la boleta agresiones" onclick="return confirm('¿Está seguro de que desea eliminar esta boleta?');"><i class="fa fa-trash fa-lg text-danger"></i></a>
                                     </td>
                                 </tr>
                             <?php } ?>
