+++ vistaoperativos.php
@@ -51,7 +51,7 @@ function myFunction() {
                                         <td><?php echo $row['fecha_hora']; ?></td>
                                         <td><?php echo $row['nombre_agente']; ?></td>
 										<td> <a href="listadooperativos.php?idFoto=<?php echo $row['id']; ?>">Ver Fotos</a></td>
-                                        <td><a href="deleteoperativos.php?id=<?php echo $row['id']; ?>" title="Click para eliminar la boleta bitacora operativos"><i class="fa fa-trash fa-lg text-danger"></i></a></td>
+                                        <td><a href="deleteoperativos.php?id=<?php echo $row['id']; ?>" title="Click para eliminar la boleta bitacora operativos" onclick="return confirm('¿Estás seguro de que deseas eliminar esta boleta?')"><i class="fa fa-trash fa-lg text-danger"></i></a></td>
                                         </tr>
 	<?php } ?>
                                     </tbody>
